backend/requirements.txt
Django==4.2.13
uvicorn[standard]==0.30.1
gunicorn==22.0.0
djangorestframework==3.15.2
django-filter==24.3
drf-yasg==1.21.7
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.0.7
django-redis==5.4.0
celery==5.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
elasticsearch==8.14.0
boto3==1.34.151
minio==7.2.6
sentry-sdk==1.45.0
prometheus-client==0.20.0
channels==4.1.0
channels-redis==4.2.0
httpx==0.26.0